moved ServerTag row to separate file + rubocop
